requirements.txt
pyautogui
mss
pynput
onnxruntime
requests
rapidocr
pillow
pytesseract
PyQt5

